+++ packages/core/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Remove BN.js in favor of native BigInt.
+- Remove BN.js in favor of native BigInt. ([#602](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/602))
+- Add support for additional network names in network manifest. ([#547](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/547))
 
 ## 1.16.1 (2022-06-16)